requirements.txt
fastapi
python-dotenv
redis
rq
python-twitter
uvicorn
